# www/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestao Projetos de Energia Solar",
        "description": "Sistema CRUD desenvolvido utilizando a framework Laravel em PHP é voltado para a gestão de projetos de energia solar. Ele adere aos princípios da arquitetura limpa (Clean Architecture), incorporando testes unitários para assegurar a qualidade do código. Além disso, oferece uma documentação abrangente para facilitar o entendimento e a manutenção do sistema.",
        "contact": {
            "email": "adriano@instalustres.com.br"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/integradores": {
            "get": {
                "tags": [
                    "Integradores"
                ],
                "summary": "Exibe lista de integradores",
                "description": "",
                "operationId": "getIntegradoresIndex",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Filtrar pelo ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "nome",
                        "in": "query",
                        "description": "Filtrar pelo nome",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Filtrar pelo email",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tipo",
                        "in": "query",
                        "description": "Filtrar pelo tipo",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "INTEGRADOR",
                                "ADMIN"
                            ]
                        }
                    },
                    {
                        "name": "ativo",
                        "in": "query",
                        "description": "Filtrar pelo status",
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ]
                        }
                    },
                    {
                        "name": "ordenar_id",
                        "in": "query",
                        "description": "Ordenar pelo ID",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "ordenar_nome",
                        "in": "query",
                        "description": "Ordenar pelo nome",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lista de integradores."
                                        },
                                        "content": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/IntegradorResource"
                                                    }
                                                },
                                                "meta": {
                                                    "$ref": "#/components/schemas/CustomPaginator"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": "200"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request."
                    },
                    "401": {
                        "description": "Unauthorized."
                    },
                    "403": {
                        "description": "Forbidden."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Integradores"
                ],
                "summary": "Cadastra novo integrador",
                "description": "",
                "operationId": "postIntegradoresStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateIntegradorRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Integrador cadastrado com sucesso."
                                        },
                                        "content": {
                                            "$ref": "#/components/schemas/IntegradorResource"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": "200"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request."
                    },
                    "401": {
                        "description": "Unauthorized."
                    },
                    "403": {
                        "description": "Forbidden."
                    },
                    "422": {
                        "description": "Unprocessable Content.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "O campo nome é obrigatório. (e mais 1 erro)"
                                        },
                                        "errors": {
                                            "properties": {
                                                "nome": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "O campo nome é obrigatório."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Envie seu e-mail e senha e retornaremos um token. Use o token no cabeçalho 'Authorization' da requisição como 'Bearer TOKEN'",
                "description": "",
                "operationId": "postIntegradoresLogin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginIntegradorRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Login efetuado com sucesso."
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "99|OOawESyfhebDhYX0LXxnn5as3fV1R3B79uZvH22Ec0df7515"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": "200"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "422": {
                        "description": "Unprocessable Content.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "O campo email é obrigatório. (e mais 1 erro)"
                                        },
                                        "errors": {
                                            "properties": {
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "O campo email é obrigatório."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/integradores/{id}/desativar": {
            "patch": {
                "tags": [
                    "Integradores"
                ],
                "summary": "Desativar o integrador",
                "description": "",
                "operationId": "pathIntegradoresDeactive",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do integrador",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 2
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Integrador desativado com sucesso."
                                        },
                                        "content": {
                                            "$ref": "#/components/schemas/IntegradorResource"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": "200"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "403": {
                        "description": "Forbidden."
                    },
                    "404": {
                        "description": "Not Found."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateIntegradorRequest": {
                "title": "Requisição para gravar integrador",
                "required": [
                    "nome",
                    "email",
                    "tipo",
                    "password",
                    "password_confirmation"
                ],
                "properties": {
                    "nome": {
                        "description": "Get the validation rules that apply to the request.",
                        "type": "string",
                        "example": "admin@admin.com"
                    },
                    "email": {
                        "type": "string",
                        "example": "admin@admin.com"
                    },
                    "tipo": {
                        "type": "string",
                        "example": "INTEGRADOR"
                    },
                    "password": {
                        "type": "string",
                        "example": "Admin@123"
                    },
                    "password_confirmation": {
                        "type": "string",
                        "example": "Admin@123"
                    }
                },
                "type": "object"
            },
            "LoginIntegradorRequest": {
                "title": "Requisição para efetuar login",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "description": "Get the validation rules that apply to the request.",
                        "type": "string",
                        "example": "admin@admin.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "Admin@123"
                    }
                },
                "type": "object"
            },
...
            "IntegradorResource": {
                "title": "Retornoadados do integrador",
                "properties": {
                    "id": {
                        "description": "Transform the resource into an array.",
                        "type": "integer",
                        "example": 1
                    },
                    "nome": {
                        "type": "string",
                        "example": "Admin"
                    },
                    "email": {
                        "type": "string",
                        "example": "admin@admin.com"
                    },
                    "tipo": {
                        "type": "string",
                        "example": "INTEGRADOR"
                    },
                    "ativo": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2024-04-30T23:29:52.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2024-04-30T23:29:52.000000Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Insira o token no formato (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}